pyproject.toml
[project]
name = "algo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "aiolimiter>=1.2.1",
    "arxiv>=2.4.0",
    "bibtexparser>=1.4.4",
    "casadi>=3.7.2",
    "cma>=4.4.2",
    "cvxpy>=1.8.1",
    "decorator>=5.2.1",
    "deepdiff>=8.6.1",
    "evotorch>=0.6.1",
    "fastmcp>=2.14.5",
    "feedparser>=6.0.12",
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "highspy>=1.13.0",
    "httpx>=0.28.1",
    "inflection>=0.5.1",
    "jupyter>=1.1.1",
    "libsvm>=3.23.0.4",
    "limiter>=0.5.0",
    "matplotlib>=3.10.7",
    "mcp>=1.26.0",
    "mlrose",
    "mlrose-hiive>=2.2.4",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "optuna>=4.7.0",
    "optunity>=1.1.1",
    "ortools>=9.15.6755",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "plotly>=6.5.2",
    "pymc>=5.27.1",
    "pyomo>=6.9.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "scikit-opt>=0.6.6",
    "scipy>=1.16.3",
    "selenium>=4.39.0",
    "statsmodels>=0.14.6",
    "streamlit>=1.51.0",
    "streamlit-explorer>=0.1.0",
    "tqdm>=4.67.1",
    "types-requests>=2.32.4.20260107",
    "typing-inspect>=0.9.0",
    "wordcloud>=1.9.4",
]
